refactor: 일기 이미지 생성 test 분리
