+++ apps/api/src/diary/spec/diary.service.spec.ts
@@ -17,7 +17,6 @@ describe('DiaryService', () => {
   let diaryService: DiaryService;
   let diaryRepository: Mock<DiaryRepository>;
   let diaryImageRepository: Mock<DiaryImageRepository>;
-  let uploadFileService: Mock<UploadFileService>;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -47,10 +46,7 @@ describe('DiaryService', () => {
         },
         {
           provide: UploadFileService,
-          useFactory: () => ({
-            getUploadedImageUrlList: jest.fn(),
-            getUploadedImageUrl: jest.fn(),
-          }),
+          useFactory: () => ({}),
         },
       ],
     }).compile();
@@ -58,89 +54,6 @@ describe('DiaryService', () => {
     diaryService = module.get<DiaryService>(DiaryService);
     diaryRepository = module.get(DiaryRepository);
     diaryImageRepository = module.get(DiaryImageRepository);
-    uploadFileService = module.get(UploadFileService);
-  });
-
-  describe('createDiaryImage', () => {
-    it('일기장 이미지 생성 성공', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      const result = await diaryService.createDiaryImage(
-        { diaryId: 1 },
-        {} as Express.Multer.File,
-        1,
-      );
-
-      expect(result).toBeUndefined();
-      expect(uploadFileService.getUploadedImageUrl).toHaveBeenCalledTimes(1);
-      expect(diaryImageRepository.create).toHaveBeenCalledTimes(1);
-    });
-
-    it('없는 일기장일 경우 NotFoundDiaryException 을 호출한다', async () => {
-      diaryRepository.findByUnique.mockResolvedValue(null);
-
-      await expect(async () => {
-        await diaryService.createDiaryImage(
-          { diaryId: 1 },
-          {} as Express.Multer.File,
-          1,
-        );
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('일기장 status 가 false 인 경우 NotFoundDiaryException 을 호출한다', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: false,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.createDiaryImage(
-          { diaryId: 1 },
-          {} as Express.Multer.File,
-          1,
-        );
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('유저의 일기장이 아닐 경우 NotUserDiaryException 을 호출한다', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        userId: 2,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.createDiaryImage(
-          { diaryId: 1 },
-          {} as Express.Multer.File,
-          1,
-        );
-      }).rejects.toThrow(new NotUserDiaryException());
-    });
   });
 
   describe('deleteDiaryImage', () => {
